+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -12,13 +12,13 @@ import lombok.Data;
 @Data
 public class User {
 
-    public User(Integer id, String name, String email) {
+    public User(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    private Integer id;
+    private Long id;
     @NotEmpty
     @NotNull
     @NotBlank
